+++ backend/apps/accounts/serializers.py
@@ -50,21 +50,6 @@ class CreateUserSerializer(serializers.ModelSerializer):
         return user
 
 
-class RegisterSerializer(serializers.ModelSerializer):
-    password = serializers.CharField(write_only=True, min_length=8)
-
-    class Meta:
-        model = User
-        fields = ["email", "name", "password"]
-
-    def create(self, validated_data):
-        password = validated_data.pop("password")
-        user = User.objects.create_user(**validated_data)
-        user.set_password(password)
-        user.save()
-        return user
-
-
 class EmailTokenObtainPairSerializer(TokenObtainPairSerializer):
     """Login local par email + mot de passe, retourne aussi le user."""